# EVG-18629 unmarshal log lines directly