[EVG-18629](https://jira.mongodb.org/browse/EVG-18629)

It appears that unmarshaling by using the decoder directly (as in [this example](https://pkg.go.dev/encoding/json#example-Decoder.Decode-Stream)) uses about half the time and an order of magnitude less memory/fewer allocations.
From inspection of the CPU profiles, (with sufficient handwaving) it could be the reason is that when we were calling the regular json.Unmarshal that called into the custom unmarshaler we were having it do double the work: the custom marshaler called back out to the json library to parse a new byte array for the inner array. At the same time the original call to the json library also needed to parse the inner array to keep track of where we were up to.

The before and after is striking:
before
```
go test -run=XXX -bench=. -benchmem  github.com/evergreen-ci/logkeeper/model
goos: darwin
goarch: arm64
pkg: github.com/evergreen-ci/logkeeper/model
BenchmarkReadLogJSONShort-10                        4248            283256 ns/op          210539 B/op       1233 allocs/op
BenchmarkReadLogJSONFewLongLines-10                    6         168763653 ns/op        139885874 B/op      2043 allocs/op
BenchmarkReadLogJSONManyShortLines-10                  5         244760442 ns/op        212711104 B/op   1200077 allocs/op
```

after
```
go test -run=XXX -bench=. -benchmem  github.com/evergreen-ci/logkeeper/model                                               
goos: darwin
goarch: arm64
pkg: github.com/evergreen-ci/logkeeper/model
BenchmarkReadLogJSONShort-10                        9276            127900 ns/op           50680 B/op        715 allocs/op
BenchmarkReadLogJSONFewLongLines-10                   14          76477333 ns/op        21030074 B/op        722 allocs/op
BenchmarkReadLogJSONManyShortLines-10                  9         120494009 ns/op        61725290 B/op     700036 allocs/op
```

It's entirely possible we could get even better, but this is definitely a good first step.