@ybrill 2022-12-28
When I looked at the heap profiles it occurred to me that the difference in memory is almost certainly because the original version reads the entire input into memory at once ([here](https://github.com/golang/go/blob/9123221ccf3c80c741ead5b6f2e960573b1676b9/src/encoding/json/stream.go#L63)) while the new version is only reading one line at a time.