@ybrill 2022-12-29
I wrote a naive parser that uses about as much memory as we do now but is many times as fast as the version in this PR.
```
go test -run=XXX -bench=. -benchmem  github.com/evergreen-ci/logkeeper/model
goos: darwin
goarch: arm64
pkg: github.com/evergreen-ci/logkeeper/model
BenchmarkReadLogJSONShort-10                               33945             34381 ns/op          118912 B/op        221 allocs/op
BenchmarkReadLogJSONFewLongLines-10                           82          13005919 ns/op        123129742 B/op       249 allocs/op
BenchmarkReadLogJSONManyShortLines-10                         43          27569861 ns/op        176579922 B/op    200071 allocs/op
BenchmarkReadLogJSONMaxLogSizeAverageLines-10                 25          46816113 ns/op        468875619 B/op     65608 allocs/op
BenchmarkReadLogJSONMaxLogSizeMaxLineSize-10                  27          40106256 ns/op        460578030 B/op        68 allocs/op
```

I can work on making it less naive if we don't mind parsing JSON ourselves in principle.