@ybrill 2023-01-03
My plan is to hold off on the naive parser and see if we need it once this PR is deployed.